refactored distributions with more detailed class hierarchy
